Added alert for when there is a match

+++ app/src/main/java/com/fiuba/tdp/linkup/services/UserService.java
@@ -1,7 +1,12 @@
 package com.fiuba.tdp.linkup.services;
 
+import android.content.Context;
+import android.content.DialogInterface;
+import android.net.ConnectivityManager;
+import android.support.v7.app.AlertDialog;
 import android.util.Log;
 
+import com.fiuba.tdp.linkup.R;
 import com.fiuba.tdp.linkup.domain.LinkUpMatch;
 import com.fiuba.tdp.linkup.domain.LinkUpUser;
 import com.fiuba.tdp.linkup.domain.LocationUser;
@@ -27,10 +32,20 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class UserService {
 
+    private final Context context;
+    Boolean endItNow;
     private UserAPIService api;
 
-    public UserService() {
-        this.api = getApi();
+    public UserService(Context context) {
+        this.context = context;
+        if (isNetworkAvailable()) {
+            Log.e("NETWOR", "AVAILABLE");
+            this.api = getApi();
+            endItNow = false;
+        } else {
+            Log.e("NETWOR", "UUUUUNAVAILABLE");
+            endItNow = true;
+        }
     }
 
     private UserAPIService getApi() {
@@ -48,6 +63,11 @@ public class UserService {
     }
 
     public void getUser(String userId, final Callback<ServerResponse<LinkUpUser>> callback) {
+        if (endItNow) {
+//            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         api.getUser(userId).enqueue(new Callback<ServerResponse<LinkUpUser>>() {
             @Override
             public void onResponse(Call<ServerResponse<LinkUpUser>> call, Response<ServerResponse<LinkUpUser>> response) {
@@ -64,9 +84,14 @@ public class UserService {
     }
 
     public void postActualFacebookUser(final String name, final String profilePicture, final Callback<ServerResponse<LinkUpUser>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         final String LOG_TAG = "POST USER";
 
-        new FacebookService().getUserData(new Callback<FacebookUserItem>() {
+        new FacebookService(context).getUserData(new Callback<FacebookUserItem>() {
             @Override
             public void onResponse(Call<FacebookUserItem> call, Response<FacebookUserItem> response) {
                 FacebookUserItem body = response.body();
@@ -117,9 +142,14 @@ public class UserService {
     }
 
     public void getUsersCompatible(final String userId, final Callback<ServerResponse<ArrayList<UserAround>>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         final String LOG_TAG = "GET AROUNDS";
 
-        api.getUsersCompatible(userId).enqueue(new Callback<ServerResponse<ArrayList<UserAround>>> () {
+        api.getUsersCompatible(userId).enqueue(new Callback<ServerResponse<ArrayList<UserAround>>>() {
             @Override
             public void onResponse(Call<ServerResponse<ArrayList<UserAround>>> call, Response<ServerResponse<ArrayList<UserAround>>> response) {
                 Log.d(LOG_TAG, "Id usuario: " + userId);
@@ -144,6 +174,10 @@ public class UserService {
 
     public void postPreferences(String userId, String gender, String distance, String minAge, String maxAge,
                                 String mode, String searchMode) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            return;
+        }
         final String LOG_TAG = "POST PREFERENCES";
 
         HashMap<String, String> parameters = new HashMap<>();
@@ -154,7 +188,7 @@ public class UserService {
         parameters.put("mode", mode);
         parameters.put("searchMode", searchMode);
 
-        Log.d(LOG_TAG,"Valores: "+userId+", "+gender+", "+distance+", "+minAge+", "+maxAge+", "+mode+", "+searchMode);
+        Log.d(LOG_TAG, "Valores: " + userId + ", " + gender + ", " + distance + ", " + minAge + ", " + maxAge + ", " + mode + ", " + searchMode);
 
         api.postPreferences(userId, parameters).enqueue(new Callback<ServerResponse>() {
             @Override
@@ -178,6 +212,11 @@ public class UserService {
 
 
     public void getUserPreferences(String userId, final Callback<ServerResponse<UserPreferences>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         api.getUserPreferences(userId).enqueue(new Callback<ServerResponse<UserPreferences>>() {
             @Override
             public void onResponse(Call<ServerResponse<UserPreferences>> call, Response<ServerResponse<UserPreferences>> response) {
@@ -195,13 +234,17 @@ public class UserService {
     }
 
     public void putLocation(String userId, LocationUser location) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            return;
+        }
         final String LOG_TAG = "POST LOCATION";
 
         HashMap<String, Double> parameters = new HashMap<>();
         parameters.put("lat", location.getLat());
         parameters.put("lon", location.getLon());
 
-        Log.d(LOG_TAG,"Valores: "+userId+", "+location.getLat()+", "+location.getLon());
+        Log.d(LOG_TAG, "Valores: " + userId + ", " + location.getLat() + ", " + location.getLon());
 
         api.putLocation(userId, parameters).enqueue(new Callback<ServerResponse>() {
             @Override
@@ -224,22 +267,74 @@ public class UserService {
     }
 
 
-
     public void updateUser(String userId, LinkUpUser user, final Callback<ServerResponse<String>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         api.updateUser(userId, user).enqueue(callback);
     }
 
     public void postLikeToUser(String myUserId, String otherUserId, final Callback<ServerResponse<Match>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
+
         HashMap<String, String> info = new HashMap<>();
         info.put("userId", otherUserId);
         api.postLikeToUser(myUserId, info).enqueue(callback);
     }
 
     public void deleteLikeToUser(String myUserId, String otherUserId, final Callback<ServerResponse<String>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         api.deleteLikeToUser(myUserId, otherUserId).enqueue(callback);
     }
 
     public void getMatchesWithoutChat(String myUserId, final Callback<ServerResponse<LinkUpMatch[]>> callback) {
+        if (endItNow) {
+            showNoConnectionAlert();
+            callback.onFailure(null, null);
+            return;
+        }
         api.getMatchesWithoutChat(myUserId).enqueue(callback);
+    }
+
+    public boolean isNetworkAvailable() {
+        final ConnectivityManager connectivityManager = ((ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE));
+        return connectivityManager.getActiveNetworkInfo() != null && connectivityManager.getActiveNetworkInfo().isConnected();
+    }
+
+    private void showNoConnectionAlert() {
+        showAlert("Atención", "No hay conexión. Por favor intenta luego");
+    }
+
+    private void showAlert(String title, String message) {
+        // 1. Instantiate an AlertDialog.Builder with its constructor
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+
+        // 2. Chain together various setter methods to set the dialog characteristics
+        builder.setMessage(message)
+                .setTitle(title);
+
+        // 3. Add the buttons
+        builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
+            public void onClick(DialogInterface dialog, int id) {
+                // User clicked OK button
+                dialog.dismiss();
+            }
+        });
+
+        // 4. Get the AlertDialog from create()
+        AlertDialog dialog = builder.create();
+
+        dialog.show();
+
     }
 }
